@@ -13,10 +13,13 @@ module LoginSupport
   end
   
   def post_login_as(user, remember_me)
-    post login_path, session: {
-      name: user.name,
-      password: user.password,
-      remember_me: remember_me,
-     }
+    post login_path,
+    params: { 
+      session: {
+        name: user.name,
+        password: user.password,
+        remember_me: remember_me,
+      }
+    }
   end
 end
